chore: move ExcludesPaths trait to illuminate/http
